Fixed #140, the shell is now an option in the CLI
