+++ boa_cli/Cargo.toml
@@ -12,3 +12,4 @@ edition = "2018"
 
 [dependencies]
 Boa = { path = "../boa" }
+structopt = "0.3.9"
